requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
google-generativeai>=0.8.0
websockets>=12.0
asyncio-mqtt>=0.16.1
python-dotenv>=1.0.0
